+++ lib/riverpie.dart
@@ -2,6 +2,7 @@ export 'package:riverpie/src/mixin.dart';
 export 'package:riverpie/src/notifier.dart';
 export 'package:riverpie/src/provider/provider.dart';
 export 'package:riverpie/src/provider/custom/future_provider.dart';
+export 'package:riverpie/src/provider/custom/state_provider.dart';
 export 'package:riverpie/src/ref.dart';
 export 'package:riverpie/src/widget/scope.dart';
 export 'package:riverpie/src/widget/consumer.dart';